requirements.txt
# this is the requirements.txt file
python-dotenv
